Replace Eigen/Dense with Eigen/Core where possible
remove unused ElementOutput classes

+++ src/visualize/Cell.cpp
@@ -32,14 +32,14 @@ NuTo::eCellTypes Cell::GetCellType() const
 
 const Eigen::VectorXd& Cell::GetData(int dataIndex) const
 {
-    if (dataIndex >= this->mData.size())
+    if (dataIndex >= static_cast<int>(this->mData.size()))
         throw VisualizeException(__PRETTY_FUNCTION__, "invalid data index.");
     return mData[dataIndex];
 }
 
 void Cell::SetData(int dataIndex, Eigen::VectorXd data)
 {
-    if (dataIndex >= this->mData.size())
+    if (dataIndex >= static_cast<int>(this->mData.size()))
         throw NuTo::VisualizeException(__PRETTY_FUNCTION__, "invalid data index.");
     mData[dataIndex] = data;
 }
